remove management and lookup services and simplified
