+++ FairwayFinder/src/FairwayFinder.Core/ConfigureCoreServices.cs
@@ -1,7 +1,5 @@
 ﻿using FairwayFinder.Core.Authorization;
 using FairwayFinder.Core.Authorization.Repositories;
-using FairwayFinder.Core.Features.CourseManagement.Repositories;
-using FairwayFinder.Core.Features.CourseManagement.Services;
 using FairwayFinder.Core.Features.Dashboard.Services;
 using FairwayFinder.Core.Features.Profile.Services;
 using FairwayFinder.Core.Features.Scorecards.Repositories;
@@ -23,12 +21,10 @@ public static class ConfigureCoreServices
     {
         // Repositories
         services.AddTransient<IUserManagementRepository, UserManagementRepository>();
-        services.AddTransient<ICourseManagementRepository, CourseManagementRepository>();
-        services.AddTransient<ICourseLookupRepository, CourseLookupRepository>();
-        services.AddTransient<ITeeboxLookupRepository, TeeboxLookupRepository>();
-        services.AddTransient<IHoleLookupRepository, HoleLookupRepository>();
+        services.AddTransient<ICourseRepository, CourseRepository>();
+        services.AddTransient<ITeeboxRepository, TeeboxRepository>();
+        services.AddTransient<IHoleRepository, HoleRepository>();
         services.AddTransient<IScorecardRepository, ScorecardRepository>();
-        services.AddTransient<IScorecardManagementRepository, ScorecardManagementRepository>();
         services.AddTransient<ILookupRepository, LookupRepository>();
         services.AddTransient<IAggregatedStatRepository, AggregatedStatRepository>();
         services.AddTransient<IScorecardStatRepository, ScorecardStatRepository>();
@@ -39,12 +35,10 @@ public static class ConfigureCoreServices
         services.AddTransient<IUserManagementService, UserManagementService>();
         services.AddTransient<IFileUploadService, UploadThingService>();
         services.AddTransient<IProfileService, ProfileService>();
-        services.AddTransient<CourseManagementService>();
-        services.AddTransient<CourseLookupService>();
-        services.AddTransient<TeeboxLookupService>();
-        services.AddTransient<HoleLookupService>();
-        services.AddTransient<ScorecardService>();
-        services.AddTransient<ScorecardManagementService>();
+        services.AddTransient<ICourseService, CourseService>();
+        services.AddTransient<ITeeboxService, TeeboxService>();
+        services.AddTransient<IHoleService, HoleService>();
+        services.AddTransient<IScorecardService, ScorecardService>();
         services.AddTransient<DashboardService>();
