+++ FairwayFinder/src/FairwayFinder.Core/Features/Course/Models/ViewModels/CourseManagementViewModel.cs
