+++ FairwayFinder/src/FairwayFinder.Core/Features/Course/Models/ViewModels/CourseViewModel.cs
