+++ FairwayFinder/src/FairwayFinder.Core/Features/Course/Models/ViewModels/TeeboxViewModel.cs
