+++ FairwayFinder/src/FairwayFinder.Core/Services/CourseService.cs
@@ -1,35 +1,53 @@
-using FairwayFinder.Core.Features.CourseManagement.Models.FormModels;
-using FairwayFinder.Core.Features.CourseManagement.Repositories;
+﻿using FairwayFinder.Core.Features.CourseManagement.Models.FormModels;
 using FairwayFinder.Core.Helpers;
 using FairwayFinder.Core.Models;
-using FairwayFinder.Core.Repositories;
-using FairwayFinder.Core.Services;
+using FairwayFinder.Core.Repositories.Interfaces;
+using FairwayFinder.Core.Services.Interfaces;
 using Microsoft.Extensions.Logging;
 
-namespace FairwayFinder.Core.Features.CourseManagement.Services;
+namespace FairwayFinder.Core.Services;
 
-public class CourseManagementService
+public class CourseService : ICourseService
 {
-    private readonly ILogger<CourseManagementService> _logger;
-    private readonly ICourseManagementRepository _courseManagementRepository;
+    private readonly ILogger<CourseService> _logger;
+    
     private readonly IUsernameRetriever _usernameRetriever;
-    private readonly CourseLookupService _courseLookupService;
-    private readonly TeeboxLookupService _teeboxLookupService;
-    private readonly HoleLookupService _holeLookupService;
+    private readonly ITeeboxService _teeboxService;
+    private readonly IHoleService _holeService;
+    
+    private readonly ICourseRepository _courseRepository;
 
-    public CourseManagementService(ICourseManagementRepository courseManagementRepository, ILogger<CourseManagementService> logger, IUsernameRetriever usernameRetriever, CourseLookupService courseLookupService, TeeboxLookupService teeboxLookupService, HoleLookupService holeLookupService)
+    public CourseService(ILogger<CourseService> logger, ICourseRepository courseRepository, IUsernameRetriever usernameRetriever, IHoleService holeService, ITeeboxService teeboxService)
     {
-        _courseManagementRepository = courseManagementRepository;
         _logger = logger;
+        _courseRepository = courseRepository;
         _usernameRetriever = usernameRetriever;
-        _courseLookupService = courseLookupService;
-        _teeboxLookupService = teeboxLookupService;
-        _holeLookupService = holeLookupService;
+        _holeService = holeService;
+        _teeboxService = teeboxService;
+    }
+    
+    public async Task<List<Course>> GetAllCoursesAsync()
+    {
+        return await _courseRepository.GetAllCoursesAsync();
+    }
+    public async Task<Course?> GetCourseByIdAsync(long courseId)
+    {
+        return await _courseRepository.GetCourseByIdAsync(courseId);
+    }
+    
+    public async Task<Course?> GetCourseByNameAsync(string name)
+    {
+        return await _courseRepository.GetCourseByNameAsync(name);
     }
 
+    public async Task<List<Course>> CourseSearchByName(string name)
+    {
+        return await _courseRepository.SearchForCourseByNameAsync(name);
+    }
+    
     public async Task<int> AddCourseAsync(CourseFormModel form)
     {
-        var course_with_name = await _courseLookupService.GetCourseByNameAsync(form.name);
+        var course_with_name = await GetCourseByNameAsync(form.name);
 
         if (course_with_name is not null)
         {
@@ -39,12 +57,12 @@ public class CourseManagementService
         
         var course = form.ToModel(new Course());
         course = EntityMetadataHelper.NewRecord(course, _usernameRetriever.UserId);
-        return await _courseManagementRepository.Insert(course);
+        return await _courseRepository.Insert(course);
     }
     
     public async Task<bool> UpdateCourseAsync(long courseId, CourseFormModel form)
     {
-        var course = await _courseLookupService.GetCourseByIdAsync(courseId);
+        var course = await GetCourseByIdAsync(courseId);
         
         // should never happen here but here in case
         if (course == null) return false;
@@ -54,12 +72,12 @@ public class CourseManagementService
         course.phone_number = form.phone_number;
         
         course = EntityMetadataHelper.UpdateRecord(course, _usernameRetriever.UserId);
-        return await _courseManagementRepository.Update(course);
+        return await _courseRepository.Update(course);
     }
 
     public async Task<int> AddTeeAsync(long courseId, TeeboxFormModel form)
     {
-        var course = await _courseLookupService.GetCourseByIdAsync(courseId);
+        var course = await GetCourseByIdAsync(courseId);
 
         if (course is null)
         {
@@ -70,13 +88,13 @@ public class CourseManagementService
         var holes = form.Holes.Select(hole => EntityMetadataHelper.NewRecord(hole.ToModel(new Hole()), _usernameRetriever.UserId)).ToList();
         tee = EntityMetadataHelper.NewRecord(tee, _usernameRetriever.UserId);
         
-        return await _courseManagementRepository.InsertNewTeeAsync(tee, holes);
+        return await _courseRepository.InsertNewTeeAsync(tee, holes);
     }
     
     
     public async Task<bool> UpdateTeeAsync(long teeboxId, TeeboxFormModel form)
     {
-        var tee = await _teeboxLookupService.GetTeeByIdAsync(teeboxId);
+        var tee = await _teeboxService.GetTeeByIdAsync(teeboxId);
 
         if (tee is null)
         {
@@ -84,7 +102,7 @@ public class CourseManagementService
             return false;
         }
         
-        var holes = await _holeLookupService.GetHolesForTeeAsync(teeboxId);
+        var holes = await _holeService.GetHolesForTeeAsync(teeboxId);
 
         for (var i = 0; i < holes.Count; i++)
         {
@@ -94,6 +112,6 @@ public class CourseManagementService
         tee = form.ToModel(tee);
         tee = EntityMetadataHelper.UpdateRecord(tee, _usernameRetriever.UserId);
         
-        return await _courseManagementRepository.UpdateTeeAsync(tee, holes);
+        return await _courseRepository.UpdateTeeAsync(tee, holes);
     }
 }
