+++ FairwayFinder/src/FairwayFinder.Core/Services/Interfaces/ITeeboxService.cs
@@ -1,0 +1,10 @@
+using FairwayFinder.Core.Models;
+
+namespace FairwayFinder.Core.Services.Interfaces;
+
+public interface ITeeboxService
+{
+    Task<Teebox?> GetTeeByIdAsync(long teeboxId);
+    Task<List<Teebox>> GetTeesForCourseAsync(long courseId);
+    Task<Dictionary<string, string>> GetTeesDropdownForCourseAsync(long courseId);
+}
