+++ FairwayFinder/src/FairwayFinder.Web/Areas/Course/Views/Course/Index.cshtml
