+++ FairwayFinder/src/FairwayFinder.Web/Areas/Course/Views/Course/ViewCourse.cshtml
