+++ FairwayFinder/src/FairwayFinder.Web/Areas/Course/Views/Shared/_GolfCourseCard.cshtml
