+++ FairwayFinder/src/FairwayFinder.Web/Areas/Course/Views/Shared/_TeeboxScorecard.cshtml
