--- FairwayFinder/src/FairwayFinder.Web/Areas/CourseManagement/Controllers/BaseCourseManagementController.cs
@@ -1,12 +1,0 @@
-﻿using FairwayFinder.Web.Controllers;
-using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.Mvc;
-
-namespace FairwayFinder.Web.Areas.CourseManagement.Controllers;
-
-[Area("CourseManagement")]
-[Authorize(Roles = Roles.Admin)]
-public class BaseCourseManagementController : BaseAuthorizedController
-{
-    
-}
