+++ FairwayFinder/src/FairwayFinder.Web/Views/Shared/Layout/_Sidebar.cshtml
@@ -24,7 +24,7 @@
             </li>
 
             <li class="nav-item mb-auto">
-                <a class="nav-link text-body" asp-area="CourseManagement" asp-controller="CourseManagement"
+                <a class="nav-link text-body" asp-area="CourseManagement" asp-controller="Course"
                    asp-action="Index">
                     <i class="fa-solid fa-golf-ball-tee pe-2"></i> Golf Courses
                 </a>
